requirements.txt
streamlit
fastf1
matplotlib
pandas
numpy
seaborn
Pillow